Further refactoring, merging FunctionTaskBase and TaskBase.
